CAM17-Pathology/demo/requirements.txt
torch>=1.10.0
 torchvision>=0.11.0
 openslide-python>=1.1.2
 numpy>=1.21.0
 matplotlib>=3.4.0
 scikit-learn>=1.0.0
 pillow>=8.3.0
